Refactor slippage logic into useSlippage hook
